+++ test/dummy/config/routes.rb
@@ -1,4 +1,4 @@
 Rails.application.routes.draw do
-  resources :payments, only: :create
+  resources :payments, only: %i[create destroy]
   post "swish/callbacks", to: "swish/callbacks#create", as: :swish_callbacks
 end